+++ src/capfinder/ml_libs.py
@@ -7,6 +7,7 @@ import jax  # noqa
 os.environ["KERAS_BACKEND"] = "jax"
 
 import keras  # noqa
+import keras.ops as K  # noqa
 from keras import layers  # noqa
 from keras.layers import LSTM  # noqa
 from keras.layers import BatchNormalization  # noqa